fix(ci): extract TreeNode type to src/types/fileTree.ts
tsc cannot resolve named exports from .vue files, so TreeNode
must live in a .ts file. FileTree.vue re-exports it for backward
compat. useFileTree imports from the .ts file directly.
